// src/model/interfaces/docente/secciones.ts
export default interface SeccionesDocente {
    per_Id: string
    asi_Id: string
    mtr_Anio: string
    mtr_Periodo: string
    nta_Seccion: string
    nta_Nivel: string
    pEs_Id: string
    asi_Asignatura: string
    sed_Sede: string
    car_Carrera: string
    fac_Facultad: string
    car_Id: string
    sed_Id: string
  }